Issue80 refactor

+++ src/main/java/se/iths/charity_shop/repository/DonationRepository.java
@@ -9,7 +9,7 @@ import se.iths.charity_shop.entity.DonationEntity;
 @Repository
 public interface DonationRepository extends CrudRepository<DonationEntity,Long> {
     @Query("SELECT SUM(m.amount) FROM DonationEntity m")
-    double getTotalAmount();
+    int getTotalAmount();
 
 
 }
